EnglishApp/requirements.txt
asgiref==3.8.1
automium==0.2.6
automium-web==0.1.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6

Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
legacy==0.1.7
MarkupSafe==3.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
views==0.3
Werkzeug==3.0.4

Django==5.1.2
django-allauth==65.1.0
django-ckeditor==6.7.1
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
django-cors-headers==4.5.0
django_autoslug==1.9.9

psycopg2-binary==2.9.10
easy-thumbnails==2.10
pillow==11.0.0
gunicorn==23.0.0
